clean up codes and fix bugs
Ensure the application is well constructed and all the codes are neat and well formatted. Also fixed the problem that search box showed "undefined" when there is no input value. Deleted unnecessary codes regarding filteredQuery.
