--- public/js/scripts.js
@@ -1,4 +1,0 @@
-let socket = io();
-socket.on('showBanner', (msg) => {
-    document.getElementById('prompt').style.display = 'block';
-})
