+++ views/layouts/main.handlebars
@@ -15,15 +15,8 @@
     <!-- Include Bootstramp's CSS-->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous" />
+    <link href="css/clothes.css" rel="stylesheet" />
 
-    <script type="module" src="/js/search.js"></script>
-
-    <style>
-        #ac-menu-1 {
-            top: 100px;
-            min-width: 108rem;
-        }
-    </style>
 </head>
 
 <body>
@@ -40,6 +33,5 @@
     integrity="sha384-cuYeSxntonz0PPNlHhBs68uyIAVpIIOZZ5JqeqvYYIcEL727kskC66kF92t6Xl2V"
     crossorigin="anonymous"></script>
 <script src="/socket.io/socket.io.js"></script>
-<script src="/js/scripts.js"></script>
 
 </html>
