+++ views/partials/filters.handlebars
@@ -107,15 +107,3 @@
         </div>
     </div>
 </div>
-
-<style>
-    .filter {
-        width: 100%;
-    }
-
-    .filter>button {
-        width: 100%;
-    }
-</style>
-
-<script src="/js/filters.js"></script>
